Add optimizers to solver infrastructure.
Also separates solver code from attr_tuner.
