+++ hoomd/tune/__init__.py
@@ -16,5 +16,6 @@ classes. Use these to implement custom tuner operations in Python code.
 from hoomd.tune.sorter import ParticleSorter
 from hoomd.tune.balance import LoadBalancer
 from hoomd.tune.custom_tuner import CustomTuner, _InternalCustomTuner
-from hoomd.tune.attr_tuner import (ManualTuneDefinition, SolverStep,
-                                   ScaleSolver, SecantSolver)
+from hoomd.tune.attr_tuner import ManualTuneDefinition
+from hoomd.tune.solve import (GradientDescent, Optimizer, RootSolver,
+                              ScaleSolver, SecantSolver, SolverStep)
